backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-binance==1.0.19
google-generativeai==0.3.2
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
websockets==12.0
httpx==0.27.0